messaging_app/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
drf-nested-routers==0.93.4
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.3